rename rotor_effective_velocity.py
