+++ floris/simulation/turbine/rotor_velocity.py
