+++ floris/tools/floris_interface.py
@@ -27,7 +27,7 @@ from floris.simulation.turbine.turbine import (
     Ct,
     power,
 )
-from floris.simulation.turbine.rotor_effective_velocity import average_velocity, rotor_effective_velocity
+from floris.simulation.turbine.rotor_velocity import average_velocity, rotor_effective_velocity
 from floris.simulation.turbine_multi_dim import multidim_power_down_select, power_multidim
 from floris.tools.cut_plane import CutPlane
 from floris.type_dec import NDArrayFloat
