+++ tests/turbine_unit_test.py
@@ -28,7 +28,7 @@ from floris.simulation import (
     power,
     Turbine,
 )
-from floris.simulation.turbine.rotor_effective_velocity import (
+from floris.simulation.turbine.rotor_velocity import (
     rotor_velocity_tilt_correction,
     rotor_velocity_yaw_correction,
     compute_tilt_angles_for_floating_turbines,
